break up providers and components
