+++ applications/desktop/src/notebook/index.js
@@ -11,7 +11,7 @@ import NotificationSystem from "react-notification-system";
 
 import configureStore from "./store";
 
-import { Notebook } from "@nteract/core/components";
+import { Notebook } from "@nteract/core/providers";
 
 import { setNotificationSystem } from "@nteract/core/actions";
