+++ packages/core/src/index.js
@@ -5,5 +5,6 @@ import * as constants from "./constants";
 import * as middlewares from "./middlewares";
 import * as reducers from "./reducers";
 import * as components from "./components";
+import * as providers from "./providers";
 
-export { actions, constants, middlewares, reducers, components };
+export { actions, constants, middlewares, reducers, components, providers };
